Change design of exercise examples page
