+++ lib/app/core/constants.dart
@@ -6,7 +6,7 @@ const kHomeGradient = [
   Color(0xFF16232B),
 ];
 
-const kFeatureButtonGradient = [
+const kButtonGradient = [
   Color.fromARGB(255, 82, 99, 112),
   Color.fromARGB(255, 65, 80, 90),
   Color(0xFF3A4750),
